Move BlogPostTemplate to templates dir
