+++ src/templates/BlogPostTemplate/blogPostTemplate.tsx
@@ -1,9 +1,9 @@
 import React from "react";
-import ButtonBlock from "../../../components/blog/ButtonBlock";
-import ImageBlock from "../../../components/blog/ImageBlock";
-import QuoteBlock from "../../../components/blog/QuoteBlock";
-import StatsBlock from "../../../components/blog/StatsBlock";
-import TextBlock from "../../../components/blog/TextBlock";
+import ButtonBlock from "../../components/blog/ButtonBlock";
+import ImageBlock from "../../components/blog/ImageBlock";
+import QuoteBlock from "../../components/blog/QuoteBlock";
+import StatsBlock from "../../components/blog/StatsBlock";
+import TextBlock from "../../components/blog/TextBlock";
 
 type BlogPostTemplateProps = {
   topic?: string;
